This should fix the templating issues.
